Create two modules, run build successfully
